chore: Add axios and react-router dependencies
The package.json file was modified to add the axios and react-router dependencies. This change will allow the application to make HTTP requests using axios and implement routing using react-router.

A new component, `movielist.jsx`, was added to the `src/components` directory. This component will be responsible for rendering a list of movies. It imports React, useState, and axios.

The `App.jsx` file was modified to import and render the `SearchBar` component. This change will display a search bar in the application.

A new component, `searchbar.jsx`, was added to the `src/components` directory. This component will handle searching for movies using the TMDB API. It imports React and useState, and includes a function to make an API request to search for movies based on a query.

The commit message suggests that the changes made are related to adding dependencies and components for movie searching and listing functionality.